added all team data available to tabs. fixed some links
