+++ src/components/Team/TeamHeader.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import strings from 'lang';
+import { HeaderContainer, Logo, Column, TeamName, Row, TeamStatsCard } from './TeamStyled';
+
+export default generalData => (
+  <HeaderContainer loading={generalData.loading} error={generalData.error}>
+    <Logo
+      src={generalData.data.logo_url}
+      alt="Team logo"
+    />
+    <Column>
+      <TeamName>{generalData.data.name}</TeamName>
+      <Row>
+        <TeamStatsCard
+          title={strings.th_wins}
+          subtitle={<div className="textSuccess">{generalData.data.wins}</div>}
+        />
+        <TeamStatsCard
+          title={strings.th_losses}
+          subtitle={<div className="textDanger">{generalData.data.losses}</div>}
+        />
+        <TeamStatsCard
+          title={strings.th_rating}
+          subtitle={Math.floor(generalData.data.rating)}
+        />
+      </Row>
+    </Column>
+  </HeaderContainer>
+);
